+++ src/Driver/Export/ReportExport.php
@@ -3,13 +3,28 @@
 namespace MBLSolutions\Report\Driver\Export;
 
 use Illuminate\Support\Collection;
+use MBLSolutions\Report\Interfaces\ExportDriver;
 use MBLSolutions\Report\Services\BuildReportService;
 use MBLSolutions\Report\Support\Maps\ReportResultMap;
 
-abstract class ReportExport
+abstract class ReportExport implements ExportDriver
 {
     /** @var array $protectedKeys */
     protected CONST PROTECTED_KEYS = ['signature', 'driver', 'expires', 'format', 'uid'];
+
+    /**
+     * Get Export Driver Name
+     *
+     * @return string|null
+     */
+    public function getName(): ?string
+    {
+        if ($this->name) {
+            return $this->name;
+        }
+
+        return null;
+    }
 
     /**
      * Get Report Builder