+++ changelog.md
@@ -1,3 +1,8 @@
+## v2.3.0
+
++ Fire event for each job chunk completed
++ Add support for queued export in multiple formats using new export interface (\MBLSolutions\Report\Interfaces\ExportDriver)
+
 ## v2.2.0
 
 + Fix report export render to build report